+++ mayan/apps/mailer/admin.py
@@ -1,13 +1,6 @@
 from django.contrib import admin
 
-from .models import LogEntry, UserMailer
-
-
-@admin.register(LogEntry)
-class LogEntryAdmin(admin.ModelAdmin):
-    date_hierarchy = 'datetime'
-    list_display = ('datetime', 'message')
-    readonly_fields = ('datetime', 'message')
+from .models import UserMailer
 
 
 @admin.register(UserMailer)