+++ mayan/apps/sources/exceptions.py
@@ -1,6 +1,2 @@
-
 class SourceException(Exception):
-    """
-    Base sources warning
-    """
-    pass
+    """Base sources warning"""