+++ mayan/apps/sources/icons.py
@@ -6,7 +6,6 @@ icon_document_create_multiple = Icon(
 icon_document_version_upload = Icon(
     driver_name='fontawesome', symbol='upload'
 )
-icon_log = Icon(driver_name='fontawesome', symbol='exclamation-triangle')
 icon_setup_sources = Icon(driver_name='fontawesome', symbol='upload')
 icon_setup_source_check_now = Icon(driver_name='fontawesome', symbol='check')
 icon_setup_source_delete = Icon(driver_name='fontawesome', symbol='times')