+++ mayan/apps/sources/models/scanner_sources.py
@@ -98,10 +98,11 @@ class SaneScanner(InteractiveSource):
                     'command_line': ' '.join(command_line),
                     'error_message': error_message
                 }
-                self.logs.create(message=message)
+                self.error_log.create(text=message)
                 raise SourceException(message)
             else:
                 stdout_file_object.seek(0)
+                self.error_log.all().delete()
                 return stdout_file_object
 
     def get_upload_file_object(self, form_data):