+++ mayan/apps/sources/models/watch_folder_sources.py
@@ -6,6 +6,7 @@ from pathlib import Path
 from django.db import models
 from django.utils.translation import ugettext_lazy as _
 
+from ..exceptions import SourceException
 from ..literals import SOURCE_CHOICE_WATCH, SOURCE_UNCOMPRESS_CHOICE_Y
 
 from .base import IntervalBaseModel
@@ -45,8 +46,13 @@ class WatchFolderSource(IntervalBaseModel):
         verbose_name = _('Watch folder')
         verbose_name_plural = _('Watch folders')
 
-    def check_source(self, test=False):
+    def _check_source(self, test=False):
         path = Path(self.folder_path)
+        # Force testing the path and raise errors for the log
+        path.lstat()
+
+        if not path.is_dir():
+            raise SourceException('Path {} is not a directory.'.format(path))
 
         if self.include_subdirectories:
             iterator = path.rglob(pattern='*')